+++ views/controller/delete-portfolio.php
@@ -1,13 +1,17 @@
 <?php
 
-require 'database.php';
+session_start();
 
-$requete = $bdd->prepare('DELETE FROM projets WHERE id=:id');
+if(!empty($_SESSION['username'])){
+    require 'database.php';
+
+    $requete = $bdd->prepare('DELETE FROM projets WHERE id=:id');
 
 
 
-$requete->execute(array(
-    'id' => $_POST['id'],
+    $requete->execute(array(
+        'id' => $_POST['id'],
 
-));
+    ));
+}
 echo "1";
