+++ views/controller/delete.php
@@ -6,14 +6,19 @@
  * Time: 17:01
  */
 
-require 'database.php';
+session_start();
 
-$requete = $bdd->prepare('DELETE FROM messages WHERE id=:id');
+if(!empty($_SESSION['username'])){
+
+    require 'database.php';
+
+    $requete = $bdd->prepare('DELETE FROM messages WHERE id=:id');
 
 
 
-$requete->execute(array(
-    'id' => $_POST['id'],
+    $requete->execute(array(
+        'id' => $_POST['id'],
 
-));
+    ));
+}
 echo "1";
